+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Lexer
   - **_Breaking Change_**: rename `MultiKindsAction` into `MultiKindAction`.
+  - Fix: expectational lexing's post check will check the output's kind instead of action's possible kinds.
 - Parser
   - **_Breaking Change_**: Reduce package size. [#31](https://github.com/DiscreteTom/retsac/issues/31)
     - Optimize the format of serialized parser data, remove unnecessary strings.